pyproject.toml
[tool.poetry]
name = "fundgrube-notifier"
version = "0.1.0"
description = ""
authors = ["Lennart Küll"]

[tool.poetry.dependencies]
python = "^3.9"
beautifulsoup4 = "^4.11.1"
pandas = "^1.5.2"
python-dotenv = "^0.21.0"
requests = "^2.28.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
